requirements.txt
python-dotenv
black 
isort  
langchain 
langchain_google_genai
google-generativeai
langgraph
grandalf